fixed bug with the OutputWindow on Visual Studio 2013
Symptoms:
* Error Message with opening Visual Studio
* Can not open the Output tool window
* Can not build the solution or project

Details:

Problem with implementation of accessing to the pool of objects ToolWindow in VS and early initialization...
The Visual Studio has not any protection or catching of any access to non initialized object in their pool - it's a big problem for users and developers plugin :)
i.e. an exception occurs in the VS, not in plugin!

However, we can use the IVsUIShell and FindToolWindow like that:

  FindToolWindow((uint)__VSFINDTOOLWIN.FTW_fFindFirst, ref guid, out frame)

FTW_fForceCreate or FTW_fFrameOnly it's the same like simple accessing to OutputWindow... so similar fall...

but this variant also not protected and for our project we should use some lazy init. in another thread for checking - really not convenient and not safe.
So, problem should be solved if we begin in any places outside of the Initialize() and constructor... it's should be as garant that all objects ToolWindow will be initialized by VS

and strange initialization (1 working, 2 not, 3 working, 4 not... etc) will be gone... hope :)

----------
Big thanks Rafael Cossovan for complex testing!