[CI.MSBuild] Fixes for Multi-Processor Environment
In general, the logic of calling of the PRE event has been moved from onProjectStarted.
However it also has required a lot of changes with manually detecting properties for our core.

Basically this is reached with reproducing of the internal logic from Microsoft.Build.Construction/BuildEngine
well, you know about this black box, so 'as is'

+ csproj special for ci with private shell.10