Fixed `#[Build cancel = true]`
In fact it will prevent `Command "Build.Cancel" is not available` for the any context of use.

An added CancelBuildState now controls the requests in each individual action for Pre/Post events and of course when build is already in progress.

This also now will ignore actions from Pre events if they used flag of the ignoring failed build (cancel request essentially means the same failure)